CICO-66276 fixes the story defects

+++ admin/controllers/adChargeGroupsCtrl.js
@@ -88,15 +88,17 @@ admin.controller('ADChargeGroupsCtrl', ['$scope', 'ADChargeGroupsSrv', '$anchorS
     * To handle delete button in edit box and list view.
     */
 	$scope.clickedDelete = function(id) {
+		chargeGroupToDelete = _.find($scope.data.charge_groups, {
+	        value: id
+	    });
+	    if (chargeGroupToDelete.name === "Allowance") {
+	    	return;
+	    }
 		var successDeletionCallback = function() {
 			$scope.$emit('hideLoader');
 			$scope.currentClickedElement = -1;
 			// delete data from scope
-			angular.forEach($scope.data.charge_groups, function(item, index) {
-	 			if (item.value === id) {
-	 				$scope.data.charge_groups.splice(index, 1);
-	 			}
- 			});
+			$scope.data.charge_groups.splice($scope.data.charge_groups.indexOf(chargeGroupToDelete), 1);
 		};
 
 		$scope.invokeApi(ADChargeGroupsSrv.deleteItem, {'value': id }, successDeletionCallback);
